Start working on formatting
