+++ BadBuilder/Configuration/ArtifactCatalog.cs
@@ -168,37 +168,3 @@ internal sealed class ArtifactCatalog
         return [..selected.OrderBy(artifact => artifact.Priority)];
     }
 }
-
-internal sealed record GitHubReleaseSource(string Owner, string Repo, string? ReleaseTag = null, string? AssetName = null);
-
-internal sealed record ArtifactDefinition(
-    string ID,
-    string DisplayName,
-    string Description,
-    string DestinationRelativePath,
-    GitHubReleaseSource? Source,
-    IReadOnlyList<InstallOperation>? Operations = null,
-    ArtifactPriority Priority = ArtifactPriority.Homebrew,
-    string? LocalArchivePath = null);
-
-internal enum ArtifactPriority
-{
-    RockBandBlitz,
-    Exploit,
-    Bootstrap,
-    Homebrew,
-}
-
-internal sealed record InstallOperation(
-    InstallOperationKind Kind,
-    string DestinationPath,
-    string? SourcePath = null,
-    string? Contents = null);
-
-internal enum InstallOperationKind
-{
-    CopyDirectory,
-    CopyFile,
-    WriteFile,
-    RenameFile,
-}
